Multiple changes
Okay so i was buisy updating it so that everything is in english. I did
various changes and not really payed attention on what i did and saved
everything because it buided properly. However with this commit the
caculation part is proken (it wont calculate anything) and i have find
out why it is so.

This also teaches me a second thing:  always keep simple and proper
commit on every thing i do and make sure that every change can be build
properly and works as intended. Stupid me :/