Main:  properly implemented the switch statement.
It was not properly implemented. It worked but there were flaws.
But now i do have a small amount of user input checking.
But it only works for whole numbers. If i put in a decimal or a letter
or special char. It will go in some soort of loop. And i still dont know
how to debug such a thing.