Input and styling: removed a user input check.
Removed a user input check code. I want first to check if my code is
complete and that i use every bit of code properly. The styling in the
rest of the project is just a part of that.

I should have done that in 2 parts but yea i was a bit to quick again. I
really need to pay attention to that.