Added gitignore file
While exploring git for windows (yea shame on me  but for now i just try
to learn the basics of it)
I came across an option to autogenereate this file. While not fully
understanding it. It seems to be convenient. If it proves to be a
hindernace ill remove it later.